pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vichaar-core"
version = "2.0.0"
description = "Research-grade multi-agent RL environment for complex decision-making"
authors = [{name = "Hridyansh"}]
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "fastapi>=0.100.0,<1.0.0",
    "uvicorn>=0.23.0,<1.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "openai>=2.7.2",
    "httpx>=0.24.0,<1.0.0",
    "pyyaml>=6.0,<7.0",
    "gymnasium>=0.28.1",
    "openenv-core>=0.2.0",
]

[project.scripts]
server = "server.app:main"

[project.urls]
Repository = "https://huggingface.co/spaces/hridyansh/vichaar-core"

[tool.setuptools]
packages = ["agents", "configs", "core", "decision", "evaluation", "server", "tasks", "training"]
